Class has been replaced with BuildQueryStatement
This class is now useless and has been replaced with the more robust
class BuildQueryStatement. The new class will allow the creation of any
SQL SELECT statement that the program needs (since the computer app
should only be able to pull from the database and not insert SELECT
statements are all thats needed).